% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateGraph.R
\name{annotateGraph}
\alias{annotateGraph}
\title{Annotate NetBox graph}
\usage{
annotateGraph(netboxResults, edgeColors = NULL, directed = TRUE,
  linker = TRUE)
}
\arguments{
\item{netboxResults}{Output from geneConnector(). 
A list with six lists (i.e. netboxGraph, netboxCommunity, netboxOutput, 
                       nodeType, moduleMembership, neighborData)
  netboxGraph is an igraph object.
  netboxCommunity is an igraph object.
  netboxOutput is a data frame.
  nodeType is a data frame.
  moduleMembership is a data frame.
  neighborData is a data frame.}

\item{edgeColors}{A table (no header) containing hex color codes for interaction types. 
The first column is interaction type and the second column is hex color code.}

\item{directed}{TRUE or FALSE}

\item{linker}{TRUE or FALSE}
}
\value{
annotated version of netboxGraph
}
\description{
This function annotates the graph based on user input. If a table of color codes 
for interaction types is provided, then the edges will be colored accordingly by 
interaction types. If \code{"directed"} is TRUE, then the edges will be arrows 
with the same directionality as the original input network for NetBox. If 
\code{"linker"} is TRUE, then linker nodes will be shown as squares while 
non-linker nodes stay as circles.
}
\examples{
data(netbox2010)

sifNetwork<-netbox2010$network
graphReduced <- networkSimplify(sifNetwork,directed = FALSE) 

geneList<-as.character(netbox2010$geneList)

results<-geneConnector(geneList=geneList,networkGraph=graphReduced,
                      pValueAdj='BH',pValueCutoff=0.05,
                      communityMethod='lec',keepIsolatedNodes=FALSE)

netboxGraphAnnotated <- annotateGraph(netboxResults = results,
                                      edgeColors = "interaction_type.color.txt",
                                      directed = TRUE,
                                      linker = TRUE)

}
\author{
Guanlan Dong, \email{guanlan_dong@g.harvard.edu}
}
\concept{netboxr}
