% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyNumber-buildCopyNumberFreqMatrix.R
\name{buildCopyNumberFreqMatrix}
\alias{buildCopyNumberFreqMatrix}
\title{Building copy number frequency matrix}
\usage{
buildCopyNumberFreqMatrix(data, scale, breaks)
}
\arguments{
\item{data}{copy number callk matrix from GISTIC2}

\item{scale}{it is (-2,2) in default GISTIC2 convention}

\item{breaks}{five breaks}
}
\value{
a data frame of copy number frequency
}
\description{
Building copy number frequency matrix
}
\examples{
cancerType<-'KIRC'
date<-'2015_08_21' 
#mutSig2CVMat<-getMutSig2CVMat(date=date,cancerType=cancerType,workDir=tmpDir)
#dim(mutSig2CVMat)
sampleSize<-333
threshold<-0.05
#worDir<-getwd()

#copyNumberFreq<-buildCopyNumberFreqMatrix(wholeCopyNumberMat,
#              scale=c(-2,2),breaks=5)/ncol(wholeCopyNumberMat)

}
\concept{netboxr}
