% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/copyNumbermRNAcorrelation-calculateCNCallGEcorrelation.R
\name{calculateCNGEcorr}
\alias{calculateCNGEcorr}
\title{calculate correlation between copy number and gene expression}
\usage{
calculateCNGEcorr(log2CNMat, CNCallMat, log2GEMat, ampGeneList,
  delGeneList, method = "MannWhitney", average, rankTestSize, workDir,
  verbose)
}
\arguments{
\item{log2CNMat}{absolute copy number value in log2 scale}

\item{CNCallMat}{copy number call data frame}

\item{log2GEMat}{gene expression mat in log2 scale}

\item{ampGeneList}{gene list of amp copy number}

\item{delGeneList}{gene list of deleted copy number}

\item{method}{method to calculate correlation, "MannWhitney" or "Pearson"}

\item{average}{average method "mean" or "median"}

\item{rankTestSize}{min sample size to calcualte rank test}

\item{workDir}{where to store data}

\item{verbose}{output detailed information}
}
\value{
list of correlation test results
}
\description{
calculate correlation between copy number and gene expression
}
\examples{
cancerType<-"KIRC"
date<-"2015_08_21" 
#mutSig2CVMat<-getMutSig2CVMat(date=date,cancerType=cancerType,workDir=tmpDir)
#dim(mutSig2CVMat)
sampleSize<-333
threshold<-0.05
#worDir<-getwd()

#corrTable<-calculateCNGEcorr(log2CNMat,CNCallMat,log2GEMat,ampGeneList,
#delGeneList,method="MannWhitney",average=corrAverage,
#rankTestSize=minRankTestSize,workDir=filePath)

}
\concept{netboxr}
