% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDFtoList.R
\name{convertDFtoList}
\alias{convertDFtoList}
\title{Convert a data.frame to a list}
\usage{
convertDFtoList(df, listItemNameCol, listItemCol)
}
\arguments{
\item{df}{a data.frame}

\item{listItemNameCol}{the name of the column that contains the categories to be split on}

\item{listItemCol}{the name of the column that contains the entries to be added to list entries}
}
\value{
a list for the given data.frame
}
\description{
Convert a data.frame to a list
}
\examples{
#df <- results$moduleMembership
#convertDFtoList(df, 'membership', 'geneSymbol')

}
\concept{netboxr}
