% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneConnector.R
\name{geneConnector}
\alias{geneConnector}
\title{Generate sub-network mapping from a candidate gene list}
\usage{
geneConnector(
  geneList,
  networkGraph,
  directed = FALSE,
  pValueAdj = "BH",
  pValueCutoff = 0.05,
  communityMethod = "ebc",
  keepIsolatedNodes = FALSE
)
}
\arguments{
\item{geneList}{A vector containing candidate gene list}

\item{networkGraph}{An igraph network graph object}

\item{directed}{A boolean value indicating whether the input network is
directed or undirected (default = FALSE)}

\item{pValueAdj}{A string for p-value correction method c("BH", "Bonferroni")
(default = "BH")}

\item{pValueCutoff}{A numeric value of p-value cutoff for linker nodes 
(default = 0.05)}

\item{communityMethod}{A string for community detection method c("ebc","lec")
(default = "ebc")}

\item{keepIsolatedNodes}{A boolean value indicating whether to keep isolated
nodes in the netboxr result (default = FALSE)}
}
\value{
a list with four lists (i.e. netboxOutput, nodeType,
                                moduleMembership, neighborData)
  netboxGraph is igraph object.
  netboxCommunity is igraph object.
  netboxOutput is a data frame.
  nodeType is a data frame.
  moduleMembership is a data frame.
  neighborData is a data frame.
}
\description{
Generate sub-network mapping from a candidate gene list
}
\details{
Explain the parameters here
}
\examples{
data(netbox2010)

sifNetwork<-netbox2010$network
graphReduced <- networkSimplify(sifNetwork,directed = FALSE) 

geneList<-as.character(netbox2010$geneList)

results<-geneConnector(geneList=geneList,networkGraph=graphReduced,
                      pValueAdj='BH',pValueCutoff=0.05,
                      communityMethod='lec',keepIsolatedNodes=FALSE)


names(results)

plot(results$netboxGraph, layout = layout_with_fr)


write.table(results$netboxOutput,
  file = "network.sif", sep = "	",
  quote = FALSE, col.names = FALSE, row.names = FALSE
)

write.table(results$neighborData,
  file = "neighborList.txt", sep = "	",
  quote = FALSE, col.names = TRUE, row.names = FALSE
)

write.table(results$moduleMembership,
  file = "memb.ebc.txt", sep = "	",
  quote = FALSE, col.names = FALSE, row.names = FALSE
)
#
write.table(results$nodeType,
  file = "nodeType.txt", sep = "	", quote = FALSE,
  col.names = FALSE, row.names = FALSE
)
#

}
\author{
Eric Minwei Liu, \email{emliu.research@gmail.com}
}
\concept{netboxr}
