% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-getCancerType.R
\name{getCancerType}
\alias{getCancerType}
\title{Retrieve TCGA analysis summary}
\usage{
getCancerType(date = "last", cancerType = "all")
}
\arguments{
\item{date}{}

\item{cancerType}{}
}
\value{
a data frame of cancer study summary
}
\description{
Retrieve TCGA analysis summary
}
\examples{
date<-"2015_08_21"
getCancerType(date)

}
\concept{netboxr}
