% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCancerTypeSummary.R
\name{getCancerTypeSummary}
\alias{getCancerTypeSummary}
\title{Retrieve TCGA tumor type data summary}
\usage{
getCancerTypeSummary(date = "latest", dataType = "TP")
}
\arguments{
\item{date}{Default is 'latest'}

\item{dataType}{Default is 'TP' for primary tumor. [ TP,NT,all,complete ]}
}
\value{
a data frame of data summary
}
\description{
Retrieve TCGA tumor type data summary
}
\examples{
#date<-getRunDates(latest=TRUE)
#cancerTypeSummary<-getCancerTypeSummary(date=date,dataType='all')
#cancerTypeSummary

}
\concept{netboxr}
