% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusterOverrepresentation.R
\name{getClusterOverrepresentation}
\alias{getClusterOverrepresentation}
\title{Run an enrichment analysis for a set of clustered data}
\usage{
getClusterOverrepresentation(clusters, gmtFile, totalEntriesNum = NULL,
  verbose = FALSE)
}
\arguments{
\item{clusters}{a set of clusters as a list of vectors}

\item{gmtFile}{a path to a GMT file where entry IDs match up the IDs in the vectors}

\item{totalEntriesNum}{a number to be used as a total entry count for phyper,
if NULL this will automatically be calculated as the number of unique entries
across all clusters}

\item{verbose}{a boolean whether to display debugging information}
}
\value{
a data.frame with the following entries:

\itemize{
  \item{entriesInGmtNum} {TODO}
  \item{gmtEntryNum} {TODO}
}
}
\description{
Run an enrichment analysis for a set of clustered data
}
\examples{
#TODO

}
\concept{
netboxr
}
