% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-getCopyNumberList.R
\name{getCopyNumberList}
\alias{getCopyNumberList}
\title{Retrieve significant copy number list from TCGA study}
\usage{
getCopyNumberList(date, cancerType, selectedSampleId = NA, workDir,
  copyNumberFreqThreshold = 0.05, corrAverage = "median",
  minRankTestSize = 4, indepThreshold = 3, corrThreshold = 0.05,
  logFCThreshold = 0.5, verbose = TRUE)
}
\arguments{
\item{date}{Default is "latest"}

\item{cancerType}{Abbreviation in the TCGA study ["KIRC","GBM"]}

\item{selectedSampleId}{Subset of TCGA sample id}

\item{workDir}{Where to save the processed result}

\item{copyNumberFreqThreshold}{Default is 0.05}

\item{corrAverage}{Default is "median"}

\item{minRankTestSize}{Default is 4}

\item{indepThreshold}{Default is 3}

\item{corrThreshold}{Default is 0.05}

\item{logFCThreshold}{Default is 0.5}

\item{verbose}{Default is TRUE}
}
\value{
a data frame of data summary
}
\description{
Retrieve significant copy number list from TCGA study
}
\examples{
date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
selectedSampleId<-NULL
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#copyNumberList<-getCopyNumberList<-function(date,cancerType, 
 #                      selectedSampleId=NA, workDir,
 #                      copyNumberFreqThreshold=0.05,
 #                      corrAverage="median",
 #                      minRankTestSize=4, indepThreshold=3,
 #                      corrThreshold=0.05, logFCThreshold=0.5,
 #                      verbose=TRUE)

}
\concept{netboxr}
