% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-getGISTIC2CNValueMat.R
\name{getGISTIC2CNValueMat}
\alias{getGISTIC2CNValueMat}
\title{Retrieve GISTIC2 all_data_by_genes.txt file from TCGA study}
\usage{
getGISTIC2CNValueMat(date = "last", cancerType, workDir)
}
\arguments{
\item{date}{Default is 'latest'}

\item{cancerType}{Abbreviation in the TCGA study ['KIRC','GBM']}

\item{workDir}{Where to save the processed result}

\item{verbose}{output detailed information [ TRUE, FALSE ]}
}
\value{
a data frame of relative copy number value from GISTIC2 reuslt
}
\description{
Retrieve GISTIC2 all_data_by_genes.txt file from TCGA study
}
\examples{
date<-getRunDates(latest=TRUE)
cancerType<-'KIRC'
#worDir<-getwd()

#gistic2CNValueMat<-getGISTIC2CNValueMat(date,cancerType,workDir)

}
\concept{netboxr}
