% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-getGISTIC2Mat.R
\name{getGISTIC2Mat}
\alias{getGISTIC2Mat}
\title{Retrieve GISTIC2 all_thresholded.by_genes.txt file from TCGA study}
\usage{
getGISTIC2Mat(date = "last", cancerType, workDir, verbose = TRUE)
}
\arguments{
\item{date}{Default is "latest"}

\item{cancerType}{Abbreviation in the TCGA study ["KIRC","GBM"]}

\item{workDir}{Where to save the processed result}

\item{verbose}{output detailed information [ TRUE, FALSE ]}
}
\value{
a data frame of copy number call from GISTIC2 reuslt
}
\description{
Retrieve GISTIC2 all_thresholded.by_genes.txt file from TCGA study
}
\examples{
date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
#worDir<-getwd()

#gistic2CNCallMat<-getGISTIC2Mat(date,cancerType,workDir)

}
\concept{netboxr}
