% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-getMutationList.R
\name{getMutationList}
\alias{getMutationList}
\title{Retrieve significant mutation list from TCGA study}
\usage{
getMutationList(date, cancerType, selectedSampleId = NA, workDir,
  mutSig2CVthreshold = 0.1, rareMutationUpperLimit = 0.3,
  rareMutationLowerLimit = 0.1, rareMutationFreq = 0.02,
  verbose = TRUE)
}
\arguments{
\item{date}{Default is "latest"}

\item{cancerType}{Abbreviation in the TCGA study ["KIRC","GBM"]}

\item{selectedSampleId}{Subset of TCGA sample id}

\item{workDir}{Where to save the processed result}

\item{mutSig2CVthreshold}{Default is 0.1}

\item{rareMutationUpperLimit}{Default is 0.3}

\item{rareMutationLowerLimit}{Default is 0.1}

\item{verbose}{Defaukt is TRUE}

\item{rateMutationFreq}{Default is 0.02}
}
\value{
a list of data summary
}
\description{
Retrieve significant mutation list from TCGA study
}
\examples{
date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
selectedSampleID<-NULL
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#mutationList<-getMutationList(date,cancerType, selectedSampleId, workDir,

}
\concept{netboxr}
