% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-getMutationMAF.R
\name{getMutationMAF}
\alias{getMutationMAF}
\title{Retrieve annotated mutation MAF file from TCGA study}
\usage{
getMutationMAF(date = "last", cancerType = "BRCA", workDir,
  verbose = TRUE)
}
\arguments{
\item{date}{Default is "latest"}

\item{cancerType}{Abbreviation in the TCGA study ["KIRC","GBM"]}

\item{workDir}{Where to save the processed result}

\item{verbose}{}
}
\value{
a data frame of annotated mutation MAF file
}
\description{
Retrieve annotated mutation MAF file from TCGA study
}
\examples{
date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
#worDir<-getwd()

#mutationMAF<-getMutationMAF(date,cancerType,workDir)

}
\concept{netboxr}
