% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-getMutSig2CVMat.R
\name{getMutSig2CVMat}
\alias{getMutSig2CVMat}
\title{Retrieve significant mutation list from TCGA study}
\usage{
getMutSig2CVMat(date = "last", cancerType = "BRCA", workDir,
  verbose = TRUE)
}
\arguments{
\item{date}{Default is 'latest'}

\item{cancerType}{Abbreviation in the TCGA study ['KIRC','GBM']}

\item{workDir}{Where to save the processed result}

\item{selectedSampleId}{Subset of TCGA sample id}
}
\value{
a data frame of MutSig2CV result
}
\description{
Retrieve significant mutation list from TCGA study
}
\examples{
date<-getRunDates(latest=TRUE)
cancerType<-'KIRC'
#worDir<-getwd()

#mutSig2CVMat<-getMutSig2CVMat(date,cancerType,workDir)

}
\concept{netboxr}
