% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-getRNAseqV2RawCounts.R
\name{getRNASeqV2RawCounts}
\alias{getRNASeqV2RawCounts}
\title{Retrieve RNASeqV2 raw counts from TCGA study}
\usage{
getRNASeqV2RawCounts(date = "last", cancerType, workDir,
  verbose = TRUE)
}
\arguments{
\item{date}{Default is 'latest'}

\item{cancerType}{Abbreviation in the TCGA study ['KIRC','GBM']}

\item{workDir}{Where to save the processed result}

\item{verbose}{output detailed information [ TRUE, FALSE ]}
}
\value{
a data frame of copy number call from GISTIC2 reuslt
}
\description{
Retrieve RNASeqV2 raw counts from TCGA study
}
\examples{
date<-getRunDates(latest=TRUE)
cancerType<-'KIRC'
#worDir<-getwd()

#rawCountsMat<-getRNASeqV2RawCounts(date=date,cancerType=cancerType,workDir=filePath)

}
\concept{netboxr}
