% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-getRunDates.R
\name{getRunDates}
\alias{getRunDates}
\title{Retrieve TCGA analysis run date}
\usage{
getRunDates(latest = FALSE)
}
\arguments{
\item{latest}{TRUE or FALSE, if this parameter is FALSE, return all analysis date.}
}
\value{
a vector of ananlysis date
}
\description{
Retrieve TCGA analysis run date
}
\examples{
getRunDates(latest=FALSE)

}
\concept{netboxr}
