% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalNullModel.R
\name{globalNullModel}
\alias{globalNullModel}
\title{Generate global null model p-value}
\usage{
globalNullModel(
  netboxGraph,
  networkGraph,
  directed,
  iterations = 30,
  numOfGenes = NULL,
  pValueAdj = "BH",
  pValueCutoff = 0.05
)
}
\arguments{
\item{netboxGraph}{igraph network graph object. This igraph object contains
NetBox algorithm identified network from \code{geneConnector} function}

\item{networkGraph}{igraph network graph object. This igraph object contains
curated network information}

\item{directed}{boolean value indicating whether the input network is
directed or undirected (default = FALSE)}

\item{iterations}{numeric value for number of iterations}

\item{numOfGenes}{numeric value for number of genes mapped in the initial
network}

\item{pValueAdj}{string for p-value correction method c("BH", "Bonferroni")
as described in the details section (default = "BH")}

\item{pValueCutoff}{numeric value of p-value cutoff for linker nodes
(default = 0.05)}
}
\value{
a list of returned results
\itemize{
\item globalNull: data frame of global randomization results
\item globalNodesResult: data frame of global null tested results based on nodes
\item globalEdgesResult: data frame of global null tested results based on edges
}
}
\description{
Randomly select the same number of nodes in the largest connected component
of netbox result as a new gene candidate list and repeat multiple times
to produce a distribution of node size and edge numbers. This distribution
will be used to produce global p-value of netbox result based on the
node size or edge numbers of largest component in the final network result.
}
\details{
P-value correction methods include the Bonferroni correction
("bonferroni") or Benjamini & Hochberg ("BH").
}
\examples{
data(netbox2010)

sifNetwork<-netbox2010$network
graphReduced <- networkSimplify(sifNetwork,directed = FALSE) 

geneList<-as.character(netbox2010$geneList)

results<-geneConnector(geneList=geneList,networkGraph=graphReduced,
                      pValueAdj='BH',pValueCutoff=0.05,
                      communityMethod='lec',keepIsolatedNodes=FALSE)

names(results)

# Suggested 100 iterations. 
# Use 5 interations in the exampel to save running time.
# globalTest <- globalNullModel(netboxGraph=results$netboxGraph, 
#                              networkGraph=graphReduced, 
#                              iterations=5, numOfGenes = 274)
}
\author{
Eric Minwei Liu, \email{emliu.research@gmail.com}
}
\concept{netboxr}
