% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-intersectCNGEmat.R
\name{intersectCNGEMat}
\alias{intersectCNGEMat}
\title{merge copy number call, copy number value, gene expression data frame}
\usage{
intersectCNGEMat(CNValueMat, CNCallMat, GEMat, geneList = NULL,
  verbose = TRUE)
}
\arguments{
\item{CNValueMat}{GISTIC2 absolute copy number value data frame}

\item{CNCallMat}{GISTIC2 copy number call data frame}

\item{GEMat}{gene expression data frame}

\item{geneList}{gene name of interest}

\item{verbose}{output detailed information [ TRUE, FALSE ]}
}
\value{
a list of data frames
}
\description{
merge copy number call, copy number value, gene expression data frame
}
\examples{
date<-getRunDates(latest=TRUE)
cancerType<-'KIRC'
#worDir<-getwd()

#mergedData<-intersectCNGEMat(CNValueMat,CNCallMat,GEMat,geneList)

}
\concept{netboxr}
