% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprint-makeCopyNumberMatrix.R
\name{makeCopyNumberMatrix}
\alias{makeCopyNumberMatrix}
\title{Make binary mutation matrix for oncoprint}
\usage{
makeCopyNumberMatrix(CNCallMat, selectedSampleId, geneList,
  verbose = TRUE)
}
\arguments{
\item{CNCallMat}{GISTIC2 copy number call data frame}

\item{selectedSampleId}{subset of sample id}

\item{geneList}{Where to save the processed result}

\item{verbose}{}
}
\value{
a binary copy number call data frame
}
\description{
Make binary mutation matrix for oncoprint
}
\examples{
date<-"2015_08_21"
cancerType<-"KIRC"
worDir<-getwd()
#mutationMAF<-getMutationMAF(date,cancerType,workDir)

#copyNumberMat<-makeCopyNumberMatrix(CNCallMat,selectedSampleId,geneList)

}
\concept{netboxr}
