% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprint-makeMutationMatrix.R
\name{makeMutationMatrix}
\alias{makeMutationMatrix}
\title{Make binary mutation matrix for oncoprint}
\usage{
makeMutationMatrix(mutationMAF, selectedSampleId, geneList, verbose)
}
\arguments{
\item{mutationMAF}{mutation MAF data frame}

\item{selectedSampleId}{subset of sample id}

\item{geneList}{Where to save the processed result}

\item{verbose}{}
}
\value{
a binary mutation data frame
}
\description{
Make binary mutation matrix for oncoprint
}
\examples{
date<-"2015_08_21"
cancerType<-"KIRC"
worDir<-getwd()
#mutationMAF<-getMutationMAF(date,cancerType,workDir)

#mutationMat<-makeMutationMatrix(mutationMAF,selectedSampleId,geneList)

}
\concept{netboxr}
