% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprint-makeOncoprint.R
\name{makeOncoprint}
\alias{makeOncoprint}
\title{Making oncoprint}
\usage{
makeOncoprint(alterationMat, cancerType, moduleNum, workDir = NULL,
  width = 15, height = 10, verbose = TRUE)
}
\arguments{
\item{alterationMat}{data frame of alteration}

\item{cancerType}{TCGA cancer study name}

\item{moduleNum}{module number}

\item{workDir}{Path to the location of file save}

\item{width}{Default is 15 inch}

\item{height}{Default is 10 inch}

\item{verbose}{Default is TRUE}
}
\value{
null
}
\description{
Making oncoprint
}
\examples{
date<-"2015_08_21"
cancerType<-"KIRC"
worDir<-getwd()
#mutationMAF<-getMutationMAF(date,cancerType,workDir)

#alterationMatSingle<-unionMultipleFrames(alterationMatMultiple)

}
\concept{netboxr}
