% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetDocument.R
\docType{data}
\name{netbox2010}
\alias{netbox2010}
\title{network coming with PLOS ONE paper.}
\format{A data frame with 19147 nodes and 2196039 interactions:
\describe{
  \item{name}{vertex gene name}
  \item{edges}{interaction types}
  ...
}}
\source{
\url{http://www.pathwaycommons.org/pc2/downloads}
}
\usage{
netbox2010
}
\value{
a data.frame
}
\description{
Loading PC2V7_03052015_woHPRD containing 19147 nodes and 2267149 interactions.
Treated as undirected network. 
Removing multiple interactions and loops.
Returning igraph network of 19147 nodes and 2196039 interactions.
}
\concept{netboxr}
\keyword{datasets}
