% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkSimplify3.R
\name{networkSimplify3}
\alias{networkSimplify3}
\title{Simplify sif network into igraph network graph object with merged edge attributes}
\usage{
networkSimplify3(sifNetwork, directed = FALSE, useCores = 1)
}
\arguments{
\item{sifNetwork}{A file with sif network format (There are three columns 
in the file separated by tab, nodeA interactionType nodeB )}

\item{directed}{Logical, treat network as directed or undirected graph}

\item{useCores}{number of cores}
}
\value{
a igraph graph object
}
\description{
This function removes duplicated edges and loops to create an igraph graph 
object from tab delimited sif formatted network file.
}
\details{
For undirected graph, \code{networkSimplify} removes duplicated edges 
and loops to create an igraph graph object from tab delimited sif 
formatted network file.  

For directed graph, \code{networkSimplify} selects the first edge and
removes the rest duplicated edges and loops to create an igraph graph
object from tab delimited sif
formatted network file.
}
\examples{
data(netbox2010)

sifNetwork<-netbox2010$network         
graphReduced<-networkSimplify(sifNetwork,directed = FALSE)

}
\author{
Eric Minwei Liu, \email{emliu.research@gmail.com}
}
\concept{netboxr}
