% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetDocument.R
\docType{data}
\name{pathway_commons_v8_reactome}
\alias{pathway_commons_v8_reactome}
\title{Pathway Commons V8 Reactome}
\format{
A list of 354 genes and a data frame of 246590 interactions
\describe{
\item{geneList}{an example list of genes}
\item{network}{Pathway Commons V8 Reactome}
...
}
}
<<<<<<< HEAD
\format{
A list of 354 genes and a data frame of 246590 interactions
\describe{
\item{geneList}{an example list of genes}
\item{network}{Pathway Commons V8 Reactome}
...
}
}
<<<<<<< HEAD
<<<<<<< HEAD
<<<<<<< HEAD
>>>>>>> 51a5128... check
=======
>>>>>>> 51a5128... check
=======
>>>>>>> 51a5128... check
=======
>>>>>>> 51a5128... check
=======
>>>>>>> 51a5128... check
\source{
\url{https://www.pathwaycommons.org}
}
\usage{
pathway_commons_v8_reactome
}
\value{
A list of two elements.
}
\description{
Contains an example gene list and Pathway Commons V8 Reactome dataset for annotateGraph().
}
\concept{netboxr}
\keyword{datasets}
