% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetDocument.R
\docType{data}
\name{PC2V7_03052015_sif_controlsExpression}
\alias{PC2V7_03052015_sif_controlsExpression}
\title{pathwayCommons v7 data frame.}
\format{A data frame with 16196 nodes and 336001 interactions:
\describe{
  \item{PARTICIPANT_A}{vertex, gene name}
  \item{INTERACTION_TYPE}{interaction type, transformed binary interaction}
  \item{PARTICIPANT_B}{vertex, gene name}
  \item{INTERACTION_DATA_SOURCE}{interaction database source}
  ...
}}
\source{
\url{http://www.pathwaycommons.org/pc2/downloads}
}
\usage{
PC2V7_03052015_sif_controlsExpression
}
\value{
a data.frame
}
\description{
Loading PC2V7_03052015_sif containing 16196 nodes and 345026 interactions.
}
\concept{netboxr}
\keyword{datasets}
