% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyNumbermRNAcorrelation-plotCNCallGE.R
\name{plotCNCallGE}
\alias{plotCNCallGE}
\title{boxplot of gene expression counts between multiple copy number groups}
\usage{
plotCNCallGE(dat, levels, geneSymbol, verbose = TRUE)
}
\arguments{
\item{dat}{Default is 'latest'}

\item{levels}{Abbreviation in the TCGA study ['KIRC','GBM']}

\item{geneSymbol}{MutSig2CV result data frame}

\item{verbose}{output detailed information}
}
\value{
null
}
\description{
boxplot of gene expression counts between multiple copy number groups
}
\examples{
cancerType<-'KIRC'
date<-'2015_08_21' 
#mutSig2CVMat<-getMutSig2CVMat(date=date,cancerType=cancerType,workDir=tmpDir)
#dim(mutSig2CVMat)
sampleSize<-333
threshold<-0.05
#worDir<-getwd()

#plotCopyNumberFreq(dataWithPos=dataWithPos,
#          threshold=copyNumberFreqThreshold,chromSize=chromSize)

}
\concept{netboxr}
