% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyNumber-plotCopyNumberFreq.R
\name{plotCopyNumberFreq}
\alias{plotCopyNumberFreq}
\title{Make copy number frequency plot}
\usage{
plotCopyNumberFreq(dataWithPos, threshold, chromSize, verbose = TRUE)
}
\arguments{
\item{dataWithPos}{Default is 'latest'}

\item{threshold}{Abbreviation in the TCGA study ['KIRC','GBM']}

\item{chromSize}{MutSig2CV result data frame}

\item{verbose}{output detailed information}
}
\value{
null
}
\description{
Make copy number frequency plot
}
\examples{
cancerType<-'KIRC'
date<-'2015_08_21' 
#mutSig2CVMat<-getMutSig2CVMat(date=date,cancerType=cancerType,workDir=tmpDir)
#dim(mutSig2CVMat)
sampleSize<-333
threshold<-0.05
#worDir<-getwd()

#plotCopyNumberFreq(dataWithPos=dataWithPos,
#          threshold=copyNumberFreqThreshold,chromSize=chromSize)

}
\concept{netboxr}
