% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation-plotMutationFreq.R
\name{plotMutationFreq}
\alias{plotMutationFreq}
\title{Make MutSig2CV result plot}
\usage{
plotMutationFreq(cancerType, mat, sampleSize, threshold, workDir, verbose)
}
\arguments{
\item{cancerType}{Abbreviation in the TCGA study ['KIRC','GBM']}

\item{mat}{MutSig2CV result data frame}

\item{sampleSize}{}

\item{threshold}{Default is 0.05}

\item{workDir}{Where to save the processed result}

\item{verbose}{output detailed information}

\item{date}{Default is 'latest'}
}
\value{
a data frame of annotated mutation MAF file
}
\description{
Make MutSig2CV result plot
}
\examples{
cancerType<-'KIRC'
date<-'2015_08_21' 
#mutSig2CVMat<-getMutSig2CVMat(date=date,cancerType=cancerType,workDir=tmpDir)
#dim(mutSig2CVMat)
sampleSize<-333
threshold<-0.05
#worDir<-getwd()

#mutationMAF<-plotMutationFreq(cancerType,mat,sampleSize,threshold,workDir)

}
\concept{netboxr}
