% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-processCNCallMat.R
\name{processCNCallMat}
\alias{processCNCallMat}
\title{Prcoess GISTIC2 copy number call from TCGA study}
\usage{
processCNCallMat(CNCallMat, verbose = TRUE)
}
\arguments{
\item{CNCallMat}{GISTIC2 copy number call data frame}

\item{verbose}{output detailed information [ TRUE, FALSE ]}
}
\value{
a data frame of copy number call from GISTIC2 reuslt
}
\description{
Prcoess GISTIC2 copy number call from TCGA study
}
\examples{
date<-getRunDates(latest=TRUE)
cancerType<-'KIRC'
#worDir<-getwd()

#gistic2CNCallMat<-processCNCallMat(gistic2CNCallMat)

}
\concept{netboxr}
