% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-processCNValueMat.R
\name{processCNValueMat}
\alias{processCNValueMat}
\title{Prcoess GISTIC2 relative copy number value from TCGA study}
\usage{
processCNValueMat(relativeCNinGISTIC, verbose = TRUE)
}
\arguments{
\item{relativeCNinGISTIC}{GISTIC2 relative copy number value data frame}

\item{verbose}{output detailed information [ TRUE, FALSE ]}
}
\value{
a data frame of absolute copy number value from GISTIC2 reuslt
}
\description{
Prcoess GISTIC2 relative copy number value from TCGA study
}
\examples{
date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
#worDir<-getwd()

#absoluteCNValueMat<-processCNValueMat(GISTIC2CNValueMat)

}
\concept{netboxr}
