% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-processGECountsMat.R
\name{processGECountsMat}
\alias{processGECountsMat}
\title{Process RNASeqV2 raw count read using edgeR TMM method}
\usage{
processGECountsMat(gExpMat, normalization = "TMM", verbose = TRUE)
}
\arguments{
\item{gExpMat}{gene expression data frame}

\item{normalization}{Abbreviation in the TCGA study ['TMM','none']}

\item{verbose}{output detailed information [ TRUE, FALSE ]}
}
\value{
a data frame of RNASeq read counts
}
\description{
Process RNASeqV2 raw count read using edgeR TMM method
}
\examples{
date<-getRunDates(latest=TRUE)
cancerType<-'KIRC'
#worDir<-getwd()

#gExpMat<-processGECountsMat(gExpMat,normalization='TMM')

}
\concept{netboxr}
