% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-processMutationMAF.R
\name{processMutationMAF}
\alias{processMutationMAF}
\title{Pre-processing mutation maf file}
\usage{
processMutationMAF(mutationMAF, verbose = TRUE)
}
\arguments{
\item{mutationMAF}{mutation maf file}

\item{verbose}{Defaukt is TRUE}
}
\value{
a list of data summary
}
\description{
Pre-processing mutation maf file
}
\examples{
date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
selectedSampleID<-NULL
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#mutationList<-getMutationList(date,cancerType, selectedSampleId, workDir,

}
\concept{netboxr}
