% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-processRawCountsMat.R
\name{processRawCountsMat}
\alias{processRawCountsMat}
\title{Process RNASeqV2 raw count read using edgeR TMM method}
\usage{
processRawCountsMat(rawCountsMat, normalization = "TMM")
}
\arguments{
\item{rawCountsMat}{raw counts data frame from TCGA}

\item{normalization}{Abbreviation in the TCGA study ["TMM","none"]}

\item{verbose}{output detailed information [ TRUE, FALSE ]}
}
\value{
a data frame of RNASeq read counts
}
\description{
Process RNASeqV2 raw count read using edgeR TMM method
}
\examples{
date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
#worDir<-getwd()

#rawCountsMat<-processRawCountsMat(rawCountsMat,normalization="TMM")

}
\concept{netboxr}
