% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-runNetBox2.R
\name{runNetBox2}
\alias{runNetBox2}
\title{Run NetBox2 network analysis}
\usage{
runNetBox2(dataDir, cancerType, mutationList, ampGeneList, delGeneList,
  epiSilencedList, mutationFreq, ampGeneFreq, delGeneFreq, epiSilencedFreq,
  pathwayCommonsDb, directed, linkerPValThreshold,
  communityDetectionMethod, keepIsolatedNodes, verbose = TRUE)
}
\arguments{
\item{dataDir}{Where to store data}

\item{cancerType}{study name from TCGA}

\item{mutationList}{User provided mutated gene list}

\item{ampGeneList}{User provided copy number amp alteration gene list}

\item{delGeneList}{User provided copy number deleted alteration gene list}

\item{epiSilencedList}{User provided epi genetic silencing gene list}

\item{mutationFreq}{User provided mutated gene freq}

\item{ampGeneFreq}{User provided copy number amp alteration gene freq}

\item{delGeneFreq}{User provided copy number deleted alteration gene freq}

\item{epiSilencedFreq}{User provided epi genetic silencing gene freq}

\item{pathwayCommonsDb}{c('signal', 'BioGrid')}

\item{directed}{TRUE or FALSE}

\item{linkerPValThreshold}{Default is 0.05}

\item{communityDetectionMethod}{'ebc', 'lec', 'none'}

\item{keepIsolatedNodes}{TRUE or FALSE}

\item{verbose}{Default is TRUE}
}
\value{
null
}
\description{
Run NetBox2 network analysis
}
\examples{
date<-getRunDates(latest=TRUE)
cancerType<-'KIRC'
selectedSampleId<-NA
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#runNetBox2(dataDir,cancerType,
#           mutationList,ampGeneList,delGeneList,epiSilencedList,
#           mutationFreq,ampGeneFreq,delGeneFreq,epiSilencedFreq,
#           pathwayCommonsDb,directed,
#           linkerPValThreshold,communityDetectionMethod,
#           keepIsolatedNodes,verbose=TRUE)

}
\concept{netboxr}
