% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprint-unionMultipleFrames.R
\name{unionMultipleFrames}
\alias{unionMultipleFrames}
\title{Union multiple data frames for making oncoprint}
\usage{
unionMultipleFrames(mat, verbose = TRUE)
}
\arguments{
\item{mat}{List of multiple data frames}

\item{verbose}{Default is TRUE}
}
\value{
a single data frame that is union of multiple data frames
}
\description{
Union multiple data frames for making oncoprint
}
\examples{
date<-"2015_08_21"
cancerType<-"KIRC"
worDir<-getwd()
#mutationMAF<-getMutationMAF(date,cancerType,workDir)

#alterationMatSingle<-unionMultipleFrames(alterationMatMultiple)

}
\concept{netboxr}
