% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprint-unionTwoFrames.R
\name{unionTwoFrames}
\alias{unionTwoFrames}
\title{Union data frames for making oncoprint}
\usage{
unionTwoFrames(mat1, mat2, verbose = TRUE)
}
\arguments{
\item{mat1}{data frame 1}

\item{mat2}{data frame 2}

\item{verbose}{Default is TRUE}
}
\value{
a union data frame of mat1 and mat2
}
\description{
Union data frames for making oncoprint
}
\examples{
date<-'2015_08_21'
cancerType<-'KIRC'
worDir<-getwd()
#mutationMAF<-getMutationMAF(date,cancerType,workDir)

#alterationMat<-unionTwoFrames(memoMutationMat,memoCNCallMat)

}
\concept{netboxr}
