% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneConnector.R
\name{geneConnector}
\alias{geneConnector}
\title{Generate sub-network mapping from a candidate gene list}
\usage{
geneConnector(
  geneList,
  networkGraph,
  directed = FALSE,
  pValueAdj = "BH",
  pValueCutoff = 0.05,
  communityMethod = "ebc",
  keepIsolatedNodes = FALSE
)
}
\arguments{
\item{geneList}{character vector containing a list of candidate genes}

\item{networkGraph}{igraph network graph object. This igrpah object contains
curated network information}

\item{directed}{boolean value indicating whether the input network is
directed or undirected (default = FALSE)}

\item{pValueAdj}{string for p-value correction method c("BH", "Bonferroni")
as described in the details section (default = "BH")}

\item{pValueCutoff}{numeric value of p-value cutoff for linker nodes
(default = 0.05)}

\item{communityMethod}{string for community detection method c("ebc","lec")
as described in the details section (default = "ebc")}

\item{keepIsolatedNodes}{A boolean value indicating whether to keep isolated
nodes in the netboxr result (default = FALSE)}
}
\value{
a list of returned netboxr results
\itemize{
\item netboxGraph: igraph object of NetBox algorithm identified network nodes
and connections
\item netboxCommunity: igraph object of network community assignment
\item netboxOutput: data frame of NetBox algorithm identified network nodes
and connections
\item nodeType: data frame of node types ("candidate" or "linker")
in the NetBox algorithm indentified network.
\item moduleMembership: data frame of module (community) membership.
\item neighborData: data frame of information of nodes directly connected to
candidate gene nodes.
}
}
\description{
Generate sub-network mapping from a candidate gene list
}
\details{
P-value correction methods include the Bonferroni correction
("bonferroni") or Benjamini & Hochberg ("BH"). Community detection methods
include using edge betweeness score ("ebc") or using leading eigenvector
method ("lec)
}
\examples{
data(netbox2010)

sifNetwork<-netbox2010$network
graphReduced <- networkSimplify(sifNetwork,directed = FALSE) 

geneList<-as.character(netbox2010$geneList)

results<-geneConnector(geneList=geneList,networkGraph=graphReduced,
                      pValueAdj='BH',pValueCutoff=0.05,
                      communityMethod='lec',keepIsolatedNodes=FALSE)


names(results)

plot(results$netboxGraph, layout = layout_with_fr)


write.table(results$netboxOutput,
  file = "network.sif", sep = "	",
  quote = FALSE, col.names = FALSE, row.names = FALSE
)

write.table(results$neighborData,
  file = "neighborList.txt", sep = "	",
  quote = FALSE, col.names = TRUE, row.names = FALSE
)

write.table(results$moduleMembership,
  file = "memb.ebc.txt", sep = "	",
  quote = FALSE, col.names = FALSE, row.names = FALSE
)
#
write.table(results$nodeType,
  file = "nodeType.txt", sep = "	", quote = FALSE,
  col.names = FALSE, row.names = FALSE
)
#

}
\author{
Eric Minwei Liu, \email{emliu.research@gmail.com}
}
\concept{netboxr}
