% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalNullModel.R
\name{globalNullModel}
\alias{globalNullModel}
\title{Generate global null model p-value}
\usage{
globalNullModel(netboxGraph, networkGraph, directed, iterations = 30,
  numOfGenes = NULL, pValueAdj = "BH", pValueCutoff = 0.05)
}
\arguments{
\item{netboxGraph}{A vector containing candidate gene list}

\item{networkGraph}{An igraph graph object}

\item{directed}{TRUE or FALSE}

\item{iterations}{TRUE of FALSE}

\item{numOfGenes}{A numeric value}

\item{pValueAdj}{A string for p-value correction method c('BH, 'Bonferroni')}

\item{pValueCutoff}{A numeric value c(0,1)}
}
\value{
a list with four lists (i.e. netboxOutput, nodeType, moduleMembership, neighborData)
}
\description{
Randomly select the same number of nodes in the largest component of netbox
result as a new gene candidate list and repeat multiple times to produce a
distribution of node size and edge numbers. This distribution will be used
to produce global p-value of netbox result based on the node size or edge
numbers of largest component in the final network result.
}
\examples{
data(netbox2010)

sifNetwork<-netbox2010$network
graphReduced <- networkSimplify(sifNetwork,directed = FALSE) 

geneList<-as.character(netbox2010$geneList)

results<-geneConnector(geneList=geneList,networkGraph=graphReduced,
                      pValueAdj='BH',pValueCutoff=0.05,
                      communityMethod='lec',keepIsolatedNodes=FALSE)

names(results)

# Suggested 100 iterations. 
# Use 5 interations in the exampel to save running time.
# globalTest <- globalNullModel(netboxGraph=results$netboxGraph, 
#                              networkGraph=graphReduced, 
#                              iterations=5, numOfGenes = 274)
}
\author{
Eric Minwei Liu, \email{emliu.research@gmail.com}
}
\concept{netboxr}
