% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localNullModel.R
\name{localNullModel}
\alias{localNullModel}
\title{Generate local null model p-value}
\usage{
localNullModel(netboxGraph, iterations = 30)
}
\arguments{
\item{netboxGraph}{igraph network graph object. This igraph object contains
NetBox algorithm identified network from \code{geneConnector} function}

\item{iterations}{numeric value for number of iterations}
}
\value{
a list of returned results
\itemize{
\item randomModularityScore: vector of modularity scores in the iterations of
local re-wiring randomization process
\item randomMean: numeric value of mean of modularity scores in the iterations
of local re-wiring randomization process
\item randomSD: numeric value of standard deviation of modularity scores
in the iterations of local re-wiring randomization process
\item modularityScoreObs: numeric value of observed modularity score in
the NetBox algorithm identified network
\item zScore: numeric value of z-score
\item pValueObs: numeric value of observed p-value
}
}
\description{
This function keeps the number of connections of each nodes in the graph but
it rewires the partners of connections and produces  modularity score. When
it repeats multiple time, a modularity score distribution will be used to
produce netbox loacl p-value.
}
\examples{
\dontrun{data(netbox2010)

sifNetwork<-netbox2010$network
graphReduced <- networkSimplify(sifNetwork,directed = FALSE) 

geneList<-as.character(netbox2010$geneList)

results<-geneConnector(geneList=geneList,networkGraph=graphReduced,
                      pValueAdj='BH',pValueCutoff=0.05,
                      communityMethod='lec',keepIsolatedNodes=FALSE)

names(results)

# Suggested 1000 iterations. 
# Use 10 interations in the exampel to save running time. 
localTest <- localNullModel(netboxGraph=results$netboxGraph, iterations=10)
}

}
\author{
Eric Minwei Liu, \email{emliu.research@gmail.com}
}
\concept{netboxr}
