% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localNullModel.R
\name{localNullModel}
\alias{localNullModel}
\title{Generate local null model p-value}
\usage{
localNullModel(netboxGraph, iterations = 30)
}
\arguments{
\item{netboxGraph}{A vector containing candidate gene list}

\item{iterations}{TRUE of FALSE}
}
\value{
a list with four lists (i.e. netboxOutput, nodeType, moduleMembership, neighborData)
}
\description{
This function keeps the number of connections of each nodes in the graph but
it rewires the partners of connections and produces  modularity score. When
it repeats multiple time, a modularity score distribution will be used to 
produce netbox loacl p-value.
}
\examples{
data(netbox2010)

#geneList<-netbox2010$geneList 
#sifNetwork<-netbox2010$network         
#result<-geneConnector(geneList=geneList,sifNetwork=network,pValueAdj="BH",
#           pValueCutoff=0.05,communityMethod="lec",keepIsolatedNodes=FALSE)

#names(result)

}
\author{
Eric Minwei Liu, \email{emliu.research@gmail.com}
}
\concept{netboxr}
