% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetDocument.R
\docType{data}
\name{netbox2010}
\alias{netbox2010}
\title{network coming with Cerami et al. PLoS One 2010 paper.}
\format{
A data frame with 9264 nodes and 68111 interactions:
\describe{
<<<<<<< HEAD
  \item{name}{vertex gene name}
  \item{edges}{interaction types}
  ...
}}
=======
\item{name}{vertex gene name}
\item{edges}{interaction types}
...
}
}
>>>>>>> 51a5128... check
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/20169195}
}
\usage{
netbox2010
}
\value{
a data.frame
}
\description{
Loading netbox2010 containing 9264 nodes and 68111 interactions.
Treated as undirected network.
After removing multiple interactions and loops.
Returning igraph network of 9264 nodes and 68111 interactions.
}
\concept{netboxr}
\keyword{datasets}
