% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetDocument.R
\docType{data}
\name{pathway_commons_v8_reactome}
\alias{pathway_commons_v8_reactome}
\title{Pathway Commons v8 reactome}
\format{A data frame with 246590 interactions and 3 variables (participant A, interaction type, participant B)}
\source{
\url{https://www.pathwaycommons.org}
}
\usage{
pathway_commons_v8_reactome
}
\value{
a data.frame
}
\description{
Pathway Commons v8 reactome
}
\concept{netboxr}
\keyword{datasets}
